# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Human Link",
    "version": "3.0.0",
    "description": "Généré depuis les routeurs Express et les schémas zod partagés (shared/src/index.ts). Authentification par cookie httpOnly (`hl_at`). Les écritures (POST, PUT, PATCH, DELETE) exigent aussi l’en-tête `X-CSRF-Token` recopiant le cookie `hl_csrf` (obtenu par GET /api/auth/csrf). Les dates sont au format AAAA-MM-JJ."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/meetings": {
      "post": {
        "tags": [
          "meetings"
        ],
        "operationId": "post_meetings",
        "summary": "Organiser une réunion",
        "description": "Rôles autorisés : manager. Authentification requise (cookie de session).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 160
                  },
                  "dateTime": {},
                  "invitees": {
                    "minItems": 1,
                    "maxItems": 200,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-f\\d]{24}$"
                    }
                  }
                },
                "required": [
                  "title",
                  "dateTime",
                  "invitees"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "meeting": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id"
                      ],
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "meeting"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "manager"
        ]
      }
    },
    "/api/meetings/{id}": {
      "delete": {
        "tags": [
          "meetings"
        ],
        "operationId": "delete_meetings_id",
        "summary": "Annuler une réunion que j’organise",
        "description": "Rôles autorisés : manager. Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "manager"
        ]
      }
    },
    "/api/meetings/{id}/respond": {
      "post": {
        "tags": [
          "meetings"
        ],
        "operationId": "post_meetings_id_respond",
        "summary": "Répondre à une invitation",
        "description": "Authentification requise (cookie de session).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "response": {
                    "type": "string",
                    "enum": [
                      "Accepted",
                      "Rejected"
                    ]
                  }
                },
                "required": [
                  "response"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true
      }
    },
    "/api/meetings/invitable": {
      "get": {
        "tags": [
          "meetings"
        ],
        "operationId": "get_meetings_invitable",
        "summary": "Personnes invitables : mon équipe et les autres managers",
        "description": "Rôles autorisés : manager. Authentification requise (cookie de session).",
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "nom": {
                            "type": "string"
                          },
                          "prenom": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "departement": {
                            "type": "string"
                          },
                          "poste": {
                            "type": "string"
                          },
                          "photoFileId": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "nom",
                          "prenom"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "manager"
        ]
      }
    },
    "/api/meetings/invitations": {
      "get": {
        "tags": [
          "meetings"
        ],
        "operationId": "get_meetings_invitations",
        "summary": "Mes invitations",
        "description": "Authentification requise (cookie de session).",
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "dateTime": {
                            "type": "string"
                          },
                          "organizer": {
                            "anyOf": [
                              {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "nom": {
                                    "type": "string"
                                  },
                                  "prenom": {
                                    "type": "string"
                                  },
                                  "email": {
                                    "type": "string"
                                  },
                                  "departement": {
                                    "type": "string"
                                  },
                                  "poste": {
                                    "type": "string"
                                  },
                                  "photoFileId": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "nom",
                                  "prenom"
                                ],
                                "additionalProperties": false
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "dateTime",
                          "organizer",
                          "status"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true
      }
    },
    "/api/meetings/organized": {
      "get": {
        "tags": [
          "meetings"
        ],
        "operationId": "get_meetings_organized",
        "summary": "Réunions que j’organise, avec le décompte des réponses",
        "description": "Rôles autorisés : manager. Authentification requise (cookie de session).",
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "dateTime": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "invited": {
                            "type": "number"
                          },
                          "accepted": {
                            "type": "number"
                          },
                          "rejected": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "dateTime",
                          "status",
                          "invited",
                          "accepted",
                          "rejected"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié ou session expirée"
          },
          "403": {
            "description": "Rôle insuffisant"
          },
          "default": {
            "description": "Erreur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "x-documented": true,
        "x-roles": [
          "manager"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hl_at"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}